requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
numpy==1.22.3
python-dotenv==0.20.0
pandas==1.4.2
yfinance==0.1.70
sendgrid==6.9.7